第五个点AC

+++ testcase_my/tensorwolf/fur.py
@@ -2,7 +2,7 @@
 from tensorwolf.executor import *
 zeros = np.zeros
 ones = np.ones
-float32 = np.float32
+float32 = np.float64
 
 
 class Session(object):
